+++ client/components/App.jsx
@@ -2,13 +2,13 @@ import React from 'react'
 
 import ErrorMessage from './ErrorMessage'
 import LoadSubreddit from './LoadSubreddit'
-import SubredditContainer from '../containers/SubredditContainer'
+import SubredditList from './SubredditList'
 
 const App = () => (
   <div className='app'>
     <ErrorMessage />
     <LoadSubreddit />
-    <SubredditContainer />
+    <SubredditList />
   </div>
 )
