+++ client/components/SubredditList.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
+import {connect} from 'react-redux'
 
 import Post from './Post'
 
@@ -18,4 +19,12 @@ Subreddit.propTypes = {
   subreddits: PropTypes.array.isRequired
 }
 
-export default Subreddit
+const mapStateToProps = (state) => {
+  return {
+    subreddits: state.subreddits
+  }
+}
+
+export default connect(
+  mapStateToProps
+)(Subreddit)
