--- client/containers/SubredditContainer.js
@@ -1,13 +1,0 @@
-import {connect} from 'react-redux'
-
-import Subreddit from '../components/Subreddit'
-
-const mapStateToProps = (state) => {
-  return {
-    subreddits: state.subreddits
-  }
-}
-
-export default connect(
-  mapStateToProps
-)(Subreddit)
